visual-validator-mcp-server/training/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
onnx>=1.13.0
onnxruntime>=1.14.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
tqdm>=4.64.0
